Add prerender module and fix redirects module
