+++ nuxt.config.ts
@@ -1,70 +1,3 @@
-import { createDirectus, readItems, rest } from '@directus/sdk';
-import type { Schema } from './types/schema';
-
-const fetchPagePermalinks = async () => {
-	if (!process.env.DIRECTUS_URL) {
-		// eslint-disable-next-line no-console
-		console.warn('DIRECTUS_URL environment variable missing');
-		return [];
-	}
-
-	if (!process.env.DIRECTUS_TV_URL) {
-		// eslint-disable-next-line no-console
-		console.warn('DIRECTUS_TV_URL environment variable missing');
-		return [];
-	}
-
-	const directus = createDirectus<Schema>(process.env.DIRECTUS_URL).with(rest());
-	const directusTv = createDirectus(process.env.DIRECTUS_TV_URL).with(rest());
-
-	const permalinks = [];
-
-	const pages = await directus.request(
-		readItems('pages', {
-			fields: ['permalink'],
-			limit: -1,
-		}),
-	);
-
-	const resources = await directus.request(readItems('resources', { fields: ['slug', { type: ['slug'] }], limit: -1 }));
-
-	const team = await directus.request(
-		readItems('team', {
-			// Filter for core team members or members with resources so we don't render like 100 empty pages
-			filter: {
-				_or: [
-					{
-						type: {
-							_eq: 'core-team',
-						},
-					},
-					{
-						resources: {
-							_nnull: true,
-						},
-					},
-				],
-			},
-			fields: ['slug'],
-			limit: -1,
-		}),
-	);
-
-	const shows = await directusTv.request(readItems('shows', { fields: ['slug'], limit: -1 }));
-
-	const episodes = await directusTv.request(
-		readItems('episodes', { fields: ['slug', { season: ['*', { show: ['slug'] }] }], limit: -1 }),
-	);
-
-	permalinks.push(...pages.map((page) => page.permalink));
-	permalinks.push(...resources.map((resource) => `/${resource.type.slug}/${resource.slug}`));
-	permalinks.push(...team.map((member) => `/team/${member.slug}`));
-	permalinks.push(...shows.map((show) => `/tv/${show.slug}`));
-	permalinks.push(...episodes.map((ep) => `/tv/${ep.season.show.slug}/${ep.slug}`));
-
-	return permalinks;
-};
-
 export default defineNuxtConfig({
 	telemetry: false,
 	ssr: true,
@@ -109,15 +42,7 @@ export default defineNuxtConfig({
 	},
 
 	experimental: {
-		payloadExtraction: true,
-	},
-
-	hooks: {
-		async 'nitro:config'(nitroConfig) {
-			const permalinks = await fetchPagePermalinks();
-			nitroConfig.prerender?.routes?.push(...permalinks);
-			nitroConfig.prerender?.routes?.push('/rss.xml');
-		},
+		sharedPrerenderData: true,
 	},
 
 	nitro: {
@@ -154,14 +79,6 @@ export default defineNuxtConfig({
 		provider: 'directus',
 		directus: {
 			baseURL: `${process.env.DIRECTUS_URL}/assets/`,
-		},
-	},
-
-	vite: {
-		vue: {
-			script: {
-				defineModel: true,
-			},
 		},
 	},
 });
